Stuff opportunity form code into its own module
